Modularize the testing fixtures
Each test section is named based on the old GTF format
