+++ README.md
@@ -24,7 +24,7 @@ These tests assume that you have greengrass-lite installled and running.
 The tests can be run by executing the following command.
 
 ```
-pytest -q -s -v --ggTestAccount=<YOUR_AWS_ACCOUNT_NUMBER> --ggTestBucket=<YOUR_TEST_BUCKET> --ggTestRegion=<TEST_REGION> --ggTestThingGroup=<YOUR_THING_GROUP> greengrassTestSuite.py
+pytest -q -s -v --ggTestAccount=<YOUR_AWS_ACCOUNT_NUMBER> --ggTestBucket=<YOUR_TEST_BUCKET> --ggTestRegion=<TEST_REGION> --ggTestThingGroup=<YOUR_THING_GROUP>
 ```
 
 ### How to run a specific test(s)?
@@ -36,7 +36,7 @@ If you want to run a specific test (e.g. `test_Deployment_3_T3`), the command
 will be as follows:
 
 ```
-pytest -q -s -v --ggTestAccount=<YOUR_AWS_ACCOUNT_NUMBER> --ggTestBucket=<YOUR_TEST_BUCKET> --ggTestRegion=<TEST_REGION> --ggTestThingGroup=<YOUR_THING_GROUP> greengrassTestSuite.py -k "test_Deployment_3_T3"
+pytest -q -s -v --ggTestAccount=<YOUR_AWS_ACCOUNT_NUMBER> --ggTestBucket=<YOUR_TEST_BUCKET> --ggTestRegion=<TEST_REGION> --ggTestThingGroup=<YOUR_THING_GROUP> ./src/aws-greengrass-testing-deployment.py  -k "test_Deployment_3_T3"
 ```
 
 More documentation on the above option can be found
