+++ pyproject.toml
@@ -1,0 +1,5 @@
+# pyproject.toml
+[tool.pytest.ini_options]
+testpaths = [
+    "./src/*",
+]
