+++ src/aws-greengrass-testing-component.py
@@ -1,0 +1,135 @@
+import time
+import pytest
+from GGTestUtils import GGTestUtils
+from SystemInterface import SystemInterface
+
+
+@pytest.fixture(scope="function")  # Runs for each test function
+def gg_util_obj(pytestconfig):
+    # Setup an instance of the GGUtils class. It is then passed to the
+    # test functions.
+    gg_util = GGTestUtils(
+        pytestconfig.getoption("ggTestAccount"),
+        pytestconfig.getoption("ggTestBucket"),
+        pytestconfig.getoption("ggTestRegion"),
+        pytestconfig.getoption("ggTestThingGroup"),
+    )
+
+    # yield the instance of the class to the tests.
+    yield gg_util
+
+    # This section is called AFTER the test is run.
+
+    # Cleanup the artifacts, components etc.
+    gg_util.cleanup()
+
+
+@pytest.fixture(scope="function")  # Runs for each test function
+def system_interface(pytestconfig):
+    interface = SystemInterface()
+
+    # yield the instance of the class to the tests.
+    yield interface
+
+    # This section is called AFTER the test is run.
+    pass
+
+
+# As a component developer, I can create Greengrass component that works on my current platform.
+def test_Component_12_T1(gg_util_obj, system_interface):
+    # I upload component "MultiPlatform" version "1.0.0" from the local store
+    component_cloud_name = gg_util_obj.upload_component_with_version(
+        "MultiPlatform", "1.0.0")
+
+    # And  I create a deployment configuration with components and configuration
+    #   | MultiPlatform | 1.0.0 |
+    # And   I deploy the deployment configuration
+    deployment_id = gg_util_obj.create_deployment(
+        gg_util_obj.get_thing_group_arn(),
+        [component_cloud_name],
+    )["deploymentId"]
+    assert deployment_id is not None
+
+    # Then the deployment completes with SUCCEEDED within 180 seconds
+    assert (gg_util_obj.wait_for_deployment_till_timeout(
+        180, deployment_id) == "SUCCEEDED")
+
+    # And  I can check the cli to see the status of component MultiPlatform is RUNNING
+    """ GG LITE CLI DOESN"T SUPPORT THIS YET. """
+
+    # And  the MultiPlatform log eventually contains the line "Hello world!" within 20 seconds
+    assert (system_interface.monitor_journalctl_for_message(
+        "ggl." + component_cloud_name[0] + ".service",
+        "Hello world! World",
+        timeout=20) is True)
+
+
+# GC developer can create a component with recipes containing s3 artifact. GGC operator can deploy it and artifact can be run.
+def test_Component_16_T1(gg_util_obj, system_interface):
+    # I upload component "HelloWorld" version "1.0.0" from the local store
+    component_cloud_name = gg_util_obj.upload_component_with_version(
+        "HelloWorld", "1.0.0")
+
+    # Give 5 sec for cloud to calculate artifact checksum and make it "DEPLOYABLE"
+    time.sleep(5)
+
+    # When I create a deployment configuration with components
+    #        | HelloWorld | 1.0.0 |
+    # And I deploy the deployment configuration
+    deployment_id = gg_util_obj.create_deployment(
+        gg_util_obj.get_thing_group_arn(),
+        [component_cloud_name],
+    )["deploymentId"]
+    assert deployment_id is not None
+
+    # Then the deployment completes with SUCCEEDED within 120 seconds
+    assert (gg_util_obj.wait_for_deployment_till_timeout(
+        120, deployment_id) == "SUCCEEDED")
+
+    # Then I can check the cli to see the status of component HelloWorld is RUNNING
+    """ GG LITE CLI DOESN"T SUPPORT THIS YET. """
+
+    # Then the HelloWorld log contains the line "Evergreen's dev experience is great!"
+    assert (system_interface.monitor_journalctl_for_message(
+        "ggl." + component_cloud_name[0] + ".service",
+        "Evergreen's dev experience is great!",
+        timeout=20,
+    ) is True)
+
+
+# As a component developer, I expect kernel to fail the deployment if the checksum of downloaded artifacts does not match with the checksum in the recipe.
+def test_Component_27_T1(gg_util_obj, system_interface):
+    # Given I upload component "HelloWorld" version "1.0.0" from the local store
+    # And I ensure component "HelloWorld" version "1.0.0" exists on cloud within 120 seconds
+    # And kernel registered as a Thing
+    # And my device is running the evergreen-kernel
+    component_cloud_name = gg_util_obj.upload_component_with_version(
+        "HelloWorld", "1.0.0")
+
+    # Give 5 sec for cloud to calculate artifact checksum and make it "DEPLOYABLE"
+    time.sleep(5)
+
+    # When I corrupt the contents of the component HelloWorld version 1.0.0 in the S3 bucket
+    assert gg_util_obj.upload_corrupt_artifacts_to_s3("HelloWorld",
+                                                      "1.0.0") is True
+
+    # When I create a deployment configuration with components
+    #        | HelloWorld | 1.0.0 |
+    # And I deploy the deployment configuration
+    deployment_id = gg_util_obj.create_deployment(
+        gg_util_obj.get_thing_group_arn(),
+        [component_cloud_name],
+    )["deploymentId"]
+    assert deployment_id is not None
+
+    # Greengrass retries 10 times with a 1 minute interval
+    # Then the deployment completes with FAILED within 630 seconds
+    assert gg_util_obj.wait_for_deployment_till_timeout(
+        630, deployment_id) == "FAILED"
+
+    # the greengrass log eventually contains the line "Failed to verify digest." within 30 seconds
+    assert (system_interface.monitor_journalctl_for_message(
+        "ggl.core.ggdeploymentd.service",
+        "Failed to verify digest.",
+        timeout=30,
+    ) is True)
