+++ src/aws-greengrass-testing-fleet-status.py
@@ -1,0 +1,64 @@
+import time
+import pytest
+from GGTestUtils import GGTestUtils
+from SystemInterface import SystemInterface
+
+
+@pytest.fixture(scope="function")  # Runs for each test function
+def gg_util_obj(pytestconfig):
+    # Setup an instance of the GGUtils class. It is then passed to the
+    # test functions.
+    gg_util = GGTestUtils(
+        pytestconfig.getoption("ggTestAccount"),
+        pytestconfig.getoption("ggTestBucket"),
+        pytestconfig.getoption("ggTestRegion"),
+        pytestconfig.getoption("ggTestThingGroup"),
+    )
+
+    # yield the instance of the class to the tests.
+    yield gg_util
+
+    # This section is called AFTER the test is run.
+
+    # Cleanup the artifacts, components etc.
+    gg_util.cleanup()
+
+
+@pytest.fixture(scope="function")  # Runs for each test function
+def system_interface(pytestconfig):
+    interface = SystemInterface()
+
+    # yield the instance of the class to the tests.
+    yield interface
+
+    # This section is called AFTER the test is run.
+    pass
+
+
+# Scenario: FleetStatus-1-T1: As a customer I can get thing information with components whose statuses have changed after an IoT Jobs deployment succeeds
+def test_FleetStatus_1_T1(gg_util_obj):
+    # When I upload component "HelloWorld" version "1.0.0" from the local store
+    component_cloud_name = gg_util_obj.upload_component_with_version(
+        "HelloWorld", "1.0.0")
+
+    # Give 5 sec for cloud to calculate artifact checksum and make it "DEPLOYABLE"
+    time.sleep(5)
+
+    # When I create a deployment configuration for deployment FirstDeployment and thing group FssThingGroup with components
+    #        | HelloWorld | 1.0.0 |
+    # And I deploy the configuration for deployment FirstDeployment
+    deployment_id = gg_util_obj.create_deployment(
+        gg_util_obj.get_thing_group_arn(),
+        [component_cloud_name],
+        "FirstDeployment",
+    )["deploymentId"]
+    assert deployment_id is not None
+
+    # Then the deployment FirstDeployment completes with SUCCEEDED within 180 seconds
+    assert (gg_util_obj.wait_for_deployment_till_timeout(
+        180, deployment_id) == "SUCCEEDED")
+
+    # And I can get the thing status as "HEALTHY" with all uploaded components within 60 seconds with groups
+    #      | FssThingGroup |
+    assert (gg_util_obj.get_ggcore_device_status(
+        60, f"{gg_util_obj.get_thing_group()}") == "HEALTHY")
